fix for Scala 3
